Rename EditSchemaAction to EditSchemaPage
